+++ Extension/src/pages/filtering-log/components/FilteringLog/FilteringLog.tsx
@@ -36,6 +36,7 @@ import { NotifierType } from '../../../../common/constants';
 import { useAppearanceTheme } from '../../../common/hooks/useAppearanceTheme';
 import { FilteringEvents } from '../FilteringEvents';
 import { Icons } from '../ui/Icons';
+import { PreserveLogModal } from '../PreserveLogModal/PreserveLogModal';
 
 import '../../styles/styles.pcss';
 
@@ -185,6 +186,7 @@ const FilteringLog = observer(() => {
 
     return (
         <>
+            {logStore.isPreserveLogModalOpen && <PreserveLogModal />}
             <CommonIcons />
             <Icons />
             {wizardStore.isModalOpen