+++ CHANGELOG.md
@@ -32,7 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated [@adguard/logger] to v2.0.0.
 - Updated [@adguard/scriptlets] to v2.2.8.
 - Updated [@adguard/tsurlfilter] to v3.4.5.
-- Updated [@adguard/tswebextension] to v3.2.8.
+- Updated [@adguard/tswebextension] to v3.2.10.
 
 ### Removed