+++ CHANGELOG.md
@@ -19,8 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Early save reactivation and missing exit prompt during active User Rules/Allowlist save [#3151].
 - Some requests are blocked in 'Inverted allowlist' mode even though there're no websites added to the Allowlist [#3193].
 - Do not enable any filters on settings import if `enabled-filters` is empty [#3136].
-- Preserve logs does not retain all records in the filtering log [#3148].
 - Settings applied notification appearing before settings are actually applied [#3278].
+- Preserve logs does not retain all records in the filtering log [#3148].
 
 
 ### Changed
@@ -34,8 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2934]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2934
 [#3193]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3193
 [#3136]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3136
-[#3148]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3148
 [#3278]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3278
+[#3148]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3148
+
 
 ## Unreleased v5.2