+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased <!-- release/v5.3 -->
+## Unreleased 5.3
 
 ### Added
 
 - Beautification to exported settings JSON for improved readability [#3069].
+- Ability to keep filtering log records without size limitations when "Preserve log" is enabled [#3148].
 
 ### Fixed
 
@@ -18,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Early save reactivation and missing exit prompt during active User Rules/Allowlist save [#3151].
 - Some requests are blocked in 'Inverted allowlist' mode even though there're no websites added to the Allowlist [#3193].
 - Do not enable any filters on settings import if `enabled-filters` is empty [#3136].
+- Settings applied notification appearing before settings are actually applied [#3278].
+- Preserve logs does not retain all records in the filtering log [#3148].
+
 
 ### Changed
 
@@ -30,8 +34,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2934]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2934
 [#3193]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3193
 [#3136]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3136
+[#3278]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3278
+[#3148]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3148
 
-## Unreleased v5.2
+
+## 5.2
 
 ### Added